feat: clean up play page conventions
