+++ src/hooks/usePlayController/index.ts
@@ -1,1 +1,1 @@
-export * from './usePlayMaiaController'
+export * from './usePlayController'
