+++ src/pages/train.tsx
@@ -334,7 +334,7 @@ const Train: React.FC<Props> = ({
               game={trainingGame}
               currentNode={trainingController.currentNode}
               orientation={trainingController.orientation}
-              moves={moves}
+              availableMoves={moves}
               move={move}
               setCurrentMove={setAndSaveCurrentMove}
               setCurrentSquare={setCurrentSquare}
@@ -457,7 +457,7 @@ const Train: React.FC<Props> = ({
               game={trainingGame}
               currentNode={trainingController.currentNode}
               orientation={trainingController.orientation}
-              moves={moves}
+              availableMoves={moves}
               move={move}
               setCurrentMove={setAndSaveCurrentMove}
               setCurrentSquare={setCurrentSquare}
